RegisterPage udah jadi
udah nambahin dependencies Google Fonts juga
