+++ lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:inventory_app/page/registerPageDosen.dart';
-import 'package:inventory_app/page/registerPageMahasiswa.dart';
+import 'package:inventory_app/page/registerPage.dart';
 
 void main() {
   runApp(MyApp());
@@ -13,23 +12,7 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       title: 'Invent APP',
-
-      home: RegisterPageDosen(),
-
+      home: RegisterPage(),
     );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({Key? key}) : super(key: key);
-
-  @override
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold();
   }
 }
