--- lib/page/registerPageDosen.dart
@@ -1,358 +1,0 @@
-//import 'dart:js';
-
-// ignore_for_file: deprecated_member_use
-
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-
-class RegisterPageDosen extends StatefulWidget {
-  const RegisterPageDosen({Key? key}) : super(key: key);
-
-  @override
-  _RegisterPageDosenState createState() => _RegisterPageDosenState();
-}
-
-class _RegisterPageDosenState extends State<RegisterPageDosen> {
-  // Kumpulan widget
-
-  Widget buildName() {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        Container(
-          margin: EdgeInsets.fromLTRB(15, 35, 15, 0),
-          alignment: Alignment.centerLeft,
-          decoration: BoxDecoration(
-            color: Colors.white,
-            borderRadius: BorderRadius.circular(23),
-            boxShadow: [
-              BoxShadow(
-                color: Colors.black26,
-                blurRadius: 6,
-                offset: Offset(0, 0),
-              )
-            ],
-          ),
-          height: 1.2 * (MediaQuery.of(context).size.height / 20),
-          child: TextField(
-            keyboardType: TextInputType.name,
-            style: TextStyle(
-              color: Colors.black87,
-            ),
-            decoration: InputDecoration(
-                border: InputBorder.none,
-                contentPadding: EdgeInsets.only(left: 20, bottom: 5),
-                hintText: 'Nama',
-                hintStyle: TextStyle(color: Colors.black38)),
-          ),
-        )
-      ],
-    );
-  }
-
-  Widget buildEmail() {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        SizedBox(
-          height: 10,
-        ),
-        Container(
-          margin: EdgeInsets.fromLTRB(15, 15, 15, 0),
-          alignment: Alignment.centerLeft,
-          decoration: BoxDecoration(
-            color: Colors.white,
-            borderRadius: BorderRadius.circular(23),
-            boxShadow: [
-              BoxShadow(
-                color: Colors.black26,
-                blurRadius: 6,
-                offset: Offset(0, 0),
-              )
-            ],
-          ),
-          height: 1.2 * (MediaQuery.of(context).size.height / 20),
-          child: TextField(
-            keyboardType: TextInputType.emailAddress,
-            style: TextStyle(
-              color: Colors.black87,
-            ),
-            decoration: InputDecoration(
-                border: InputBorder.none,
-                contentPadding: EdgeInsets.only(left: 20, bottom: 5),
-                hintText: 'Email',
-                hintStyle: TextStyle(color: Colors.black38)),
-          ),
-        )
-      ],
-    );
-  }
-
-  Widget buildNIP() {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        SizedBox(
-          height: 10,
-        ),
-        Container(
-          margin: EdgeInsets.fromLTRB(15, 15, 15, 0),
-          alignment: Alignment.centerLeft,
-          decoration: BoxDecoration(
-            color: Colors.white,
-            borderRadius: BorderRadius.circular(23),
-            boxShadow: [
-              BoxShadow(
-                color: Colors.black26,
-                blurRadius: 6,
-                offset: Offset(0, 0),
-              )
-            ],
-          ),
-          height: 1.2 * (MediaQuery.of(context).size.height / 20),
-          child: TextField(
-            keyboardType: TextInputType.number,
-            style: TextStyle(
-              color: Colors.black87,
-            ),
-            decoration: InputDecoration(
-                border: InputBorder.none,
-                contentPadding: EdgeInsets.only(left: 20, bottom: 5),
-                hintText: 'NIP',
-                hintStyle: TextStyle(color: Colors.black38)),
-          ),
-        )
-      ],
-    );
-  }
-
-  Widget buildFakultas() {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        SizedBox(
-          height: 10,
-        ),
-        Container(
-          margin: EdgeInsets.fromLTRB(15, 15, 15, 0),
-          alignment: Alignment.centerLeft,
-          decoration: BoxDecoration(
-            color: Colors.white,
-            borderRadius: BorderRadius.circular(23),
-            boxShadow: [
-              BoxShadow(
-                color: Colors.black26,
-                blurRadius: 6,
-                offset: Offset(0, 0),
-              )
-            ],
-          ),
-          height: 1.2 * (MediaQuery.of(context).size.height / 20),
-          child: TextField(
-            keyboardType: TextInputType.emailAddress,
-            style: TextStyle(
-              color: Colors.black87,
-            ),
-            decoration: InputDecoration(
-                border: InputBorder.none,
-                contentPadding: EdgeInsets.only(left: 20, bottom: 5),
-                hintText: 'Fakultas',
-                hintStyle: TextStyle(color: Colors.black38)),
-          ),
-        )
-      ],
-    );
-  }
-
-  Widget buildPassword() {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        SizedBox(
-          height: 10,
-        ),
-        Container(
-          margin: EdgeInsets.fromLTRB(15, 15, 15, 0),
-          alignment: Alignment.centerLeft,
-          decoration: BoxDecoration(
-            color: Colors.white,
-            borderRadius: BorderRadius.circular(23),
-            boxShadow: [
-              BoxShadow(
-                color: Colors.black26,
-                blurRadius: 6,
-                offset: Offset(0, 0),
-              )
-            ],
-          ),
-          height: 1.1 * (MediaQuery.of(context).size.height / 20),
-          child: TextField(
-            keyboardType: TextInputType.visiblePassword,
-            style: TextStyle(
-              color: Colors.black87,
-            ),
-            decoration: InputDecoration(
-                border: InputBorder.none,
-                contentPadding: EdgeInsets.only(left: 20, bottom: 5),
-                hintText: 'Password',
-                hintStyle: TextStyle(color: Colors.black38)),
-          ),
-        )
-      ],
-    );
-  }
-
-  Widget buildSignUpButton() {
-    return Container(
-      margin: EdgeInsets.fromLTRB(0, 40, 20, 0),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.end,
-        children: <Widget>[
-          Container(
-            height: 1.2 * (MediaQuery.of(context).size.height / 20),
-            width: 3.5 * (MediaQuery.of(context).size.width / 10),
-            margin: EdgeInsets.only(bottom: 20),
-            child: OutlinedButton(
-              style: OutlinedButton.styleFrom(
-                primary: Colors.white,
-                backgroundColor: Color(0xFFEA340C),
-                side: BorderSide(width: 3.0, color: Colors.white),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(30.0),
-                ),
-              ),
-              onPressed: () {},
-              child: Text(
-                "Sign Up",
-                style: TextStyle(
-                  //color: Colors.white,
-                  letterSpacing: 1.5,
-                  fontSize: MediaQuery.of(context).size.height / 45,
-                ),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget buildContainer() {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: <Widget>[
-        Row(
-          mainAxisAlignment: MainAxisAlignment.start,
-          children: <Widget>[
-            SizedBox(
-              width: 10,
-            ),
-            Text(
-              "Dosen",
-              style: TextStyle(
-                fontSize: MediaQuery.of(context).size.height / 45,
-              ),
-            ),
-            SizedBox(
-              width: 40,
-            ),
-            Text(
-              "Mahasiswa",
-              style: TextStyle(
-                fontSize: MediaQuery.of(context).size.height / 45,
-              ),
-            ),
-          ],
-        ),
-        SizedBox(
-          height: 10,
-        ),
-
-        // KOTAK PEMBUNGKUS NAMA, EMAIL, NIP, FAKULTAS, PASSWORD
-        ClipRRect(
-          borderRadius: BorderRadius.all(
-            Radius.circular(15),
-          ),
-          child: Container(
-            height: MediaQuery.of(context).size.height * 0.6,
-            width: MediaQuery.of(context).size.width * 0.9,
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.all(Radius.circular(15)),
-              color: Colors.white10.withOpacity(0.3),
-              border: Border.all(
-                color: Colors.white,
-                width: 2,
-              ),
-            ),
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.start,
-              crossAxisAlignment: CrossAxisAlignment.center,
-              children: <Widget>[
-                buildName(),
-                buildEmail(),
-                buildNIP(),
-                buildFakultas(),
-                buildPassword(),
-                buildSignUpButton(),
-              ],
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: AnnotatedRegion<SystemUiOverlayStyle>(
-        value: SystemUiOverlayStyle.light,
-        child: GestureDetector(
-          child: Stack(
-            children: <Widget>[
-              Container(
-                height: MediaQuery.of(context).size.height,
-                width: MediaQuery.of(context).size.width,
-                decoration: BoxDecoration(
-                  gradient: LinearGradient(
-                    begin: Alignment.topCenter,
-                    end: Alignment.bottomCenter,
-                    colors: [
-                      Color(0xFFFF9F69),
-                      Color(0x79FFB183),
-                    ],
-                  ),
-                ),
-                child: SingleChildScrollView(
-                  physics: const BouncingScrollPhysics(),
-                  padding: EdgeInsets.symmetric(horizontal: 25, vertical: 100),
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.start,
-                    children: <Widget>[
-                      Text(
-                        'Registrasi',
-                        style: TextStyle(
-                          color: Colors.black,
-                          fontSize: 40,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                      SizedBox(
-                        height: 50,
-                      ),
-                      Column(
-                        children: <Widget>[
-                          buildContainer(),
-                        ],
-                      ),
-                    ],
-                  ),
-                ),
-              )
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
